Move VerticalScrollArea > widget/gui, _HiddenVerticalScrollArea > OWWidget
